==> cpu/notebooks_refacto/11-Calibration_broken_with_event.ipynb <==
# Calibration of a dysfunctional case considering events

This notebook shows a calibration of a dysfunctional case (a broken fan) based on model including the event simulation.

The calibration methodology is still the same as a [basic calibration on transient data](calibration_transient_basics.ipynb): the heat exchange coefficient adder is computed to fit the temperature increase during a time step.

### Model setup

from cpu.systems import CPUSystem

cpu = CPUSystem.load("data/cpu_ref.json")

### Calibration setup

import numpy as np
import pandas as pd

from cosapp.drivers import NonLinearSolver, EulerExplicit
from cosapp.drivers.time.scenario import Interpolator
from cosapp.recorders import DataFrameRecorder

time_driver = cpu.add_driver(EulerExplicit())
solver = time_driver.add_child(
    NonLinearSolver("solver", tol=1e-14, eps=1e-5, max_iter=10, factor=1.0)
)

# load the operation data, including fan breakage event
data = pd.read_csv("data/cpu_hot_day_intensive_use_broken.csv")
next_T = data["T_cpu"][1:]

# load the event detection
fan_broken_diag = pd.read_csv("data/fan_working_diagnostic.csv")
fan_diag = fan_broken_diag["fan_working_diagnostic"][:-1].astype(np.float64)

# define a calibration methodology
solver.add_equation("cpu.expected_next_T == cpu.next_T").add_unknown("exchanger.h_adder")

time_driver.time_interval = (0, 29)
time_driver.dt = 1.0

# define a simulation scenario using the operating conditions
time_driver.set_scenario(
    init={"T_cpu": 10.0},
    values={
        "fan.mass_flow_scalar": Interpolator(np.stack([np.linspace(0, 29, 30), fan_diag], axis=1)),
        "cpu.expected_next_T": Interpolator(np.stack([np.linspace(0, 29, 30), next_T], axis=1)),
        "fan.T_air": Interpolator(np.stack([data.index, data["fan.T_air"]], axis=1)),
        "cpu.usage": Interpolator(np.stack([data.index, data["cpu.usage"]], axis=1)),
    },
)

rec = time_driver.add_recorder(DataFrameRecorder(includes=["*"], hold=False), period=1.0)

cpu.run_drivers()
rec.data.to_csv("data/calibrated_simulation_broken2.csv")

### Analysis

from cpu.utils.plot_recorders import plot_recorders

plot_recorders(
    {
        "operation": pd.read_csv("data/cpu_hot_day_intensive_use_broken.csv"),
        "calibrated": pd.read_csv("data/calibrated_simulation_broken2.csv"),
    },
    [
        [("time", "fan.T_air"), ("time", "cpu.usage")],
        [("time", "fan.tension"), ("time", "T_cpu")],
        [("time", "exchanger.h_adder")],
    ],
    width=800,
    height=600,
)

The results are completely wrong because we calibrated a fully working model on a dataset including a breakage!

np.median(rec.data["exchanger.h_adder"])
